store/todo: tasks.pos is INTEGER in the substrate
Task.Pos was int, which lift's type table maps to TEXT; the ordering
spine (extra.sql's (pos, created_seq) index) would sort lexically the
day anything orders in SQL. int64 in the metadata, regenerated: ddl
emits INTEGER, the domain carries int64. Ordering today happens in Go,
so nothing regressed; the drift guard confirms byte-identity.

+++ store/todo/domain/task.go
@@ -14,7 +14,7 @@ import (
 type Task struct {
 	Id         string `db:"id"`
 	CreatedSeq int64  `db:"created_seq"`
-	Pos        int    `db:"pos"`
+	Pos        int64  `db:"pos"`
 	Status     string `db:"status"`
 	Text       string `db:"text"`
 	UpdatedSeq int64  `db:"updated_seq"`
